Nova main
